stock_analysis_app/requirements.txt
databricks-connect
databricks-sql-connector
python-dotenv
pandas
streamlit
numpy
lightweight-charts
streamlit-local-storage
